Refact : AcceptanceTest 리팩토링
###작업 내용###
- 중복 코드 제거
